Mappers de parameters completos
Se adjuntan los mappers de parameters completos para su pull, restan los mappers del Core.

+++ PackageDelivery.Repository.Implementation/Mappers/Parameters/PersonRepositoryMapper.cs
@@ -18,7 +18,7 @@ namespace PackageDelivery.Repository.Implementation.Mappers.Parameters
                 cellphone = input.telefono,
                 email = input.correo,
                 identificationNumber = input.documento,
-                IdAddress = input.idDireccion,
+                IdAddress = (int)input.idDireccion,
                 IdDocumentType = input.idTipoDocumento
             };
         }
